Moved task creating mutual dependency in tasks.py
