+++ src/tasks.py
@@ -8,10 +8,12 @@ from os.path import join
 import requests
 from buildercore import core, cfngen, config, project, bootstrap
 from fabric.api import sudo, run, local, task
+from fabric.contrib.console import confirm
 from decorators import echo_output, requires_aws_stack, requires_project, debugtask
 from buildercore.core import stack_conn
 import utils, aws
 from buildercore.decorators import osissue, osissuefn
+import bakery
 
 @debugtask
 @requires_project
@@ -44,6 +46,28 @@ def ami_for_project(pname):
 
     # good for figuring out filters 
     #print results[0].__dict__
+
+@task
+@requires_aws_stack
+def create_ami(stackname):
+    pname = core.project_name_from_stackname(stackname)
+    msg = "this will create a new AMI for the project %r. Continue?" % pname
+    if not confirm(msg, default=False):
+        print 'doing nothing'
+        return
+    amiid = bakery.create_ami(stackname)
+    #amiid = "ami-e9ff3682"
+    print 'AWS is now creating AMI with id', amiid
+    path = pname + '.aws.ami'
+    # wait until ami finished creating?
+    #core.update_project_file(pname + ".aws.ami", amiid)
+    new_project_file = project.update_project_file(path, amiid)
+    output_file = project.write_project_file(new_project_file)
+    print '\n' * 4
+    print 'wrote', output_file
+    print 'updated project file with new ami. these changes must be merged and committed manually'
+    print '\n' * 4
+
 
 #
 #
